renaming material to be consistent with other modules
