+++ packages/stripe/src/stripe.module.ts
@@ -3,7 +3,7 @@ import {NgModule} from '@angular/core'
 import {BrowserModule} from '@angular/platform-browser'
 import {CommonModule} from '@angular/common'
 import {FormsModule, ReactiveFormsModule} from '@angular/forms'
-import {MaterialModules} from '../../angular/src/material'
+import {MaterialModules} from '../../angular/src/material.module'
 import {StripePaymentMethodComponent} from './payment-method.component'
 import {StripePaymentMethodItemComponent} from './payment-method-item.component'
 import {StripePaymentMethodItemDisplayComponent} from './payment-method-item-display.component'
